Update ComputeNewAngle.m
line 48: it was checking the (angle of the ant and the pheromone). It needs to check btwn the (angle ant is facing) and (angle of ant & pheromone)

line 68: new angle would be just relativeAngle, which is (angle of ant & pheromone). ant_angle + relative Angle makes ant face like opposite direction